refactor: Auth components refactored, since /login and /register pages feature same layout
