+++ src/app/(auth)/login/page.tsx
@@ -1,35 +1,13 @@
-import {
-  OAuthLogin,
-  CredentialsLogin,
-  AvailableOAuthProviders,
-} from "@/components/Auth";
+import { Auth } from "@/components/Auth";
 
 export default function Page() {
   return (
-    <main className="grid grid-cols-2 gap-4 lg:grid-cols-1 lg:flex-col">
-      <img
-        aria-hidden
-        src="/slides/slide-signin.png"
-        alt="Sign in"
-        className="aspect-square w-full object-cover"
-      />
-      <section className="grid flex-1 place-content-center p-4">
-        {AvailableOAuthProviders.length > 0 && (
-          <>
-            <div className="grid gap-3">
-              {AvailableOAuthProviders.map(provider => (
-                <OAuthLogin key={provider} name={provider} />
-              ))}
-            </div>
-            <div className="relative inset-x-0 my-8 h-[2px] rounded-md bg-[#666] bg-opacity-25 uppercase">
-              <span className="absolute left-1/2 top-1/2 z-10 -translate-x-1/2 -translate-y-1/2 rounded-md bg-white px-6 py-1 font-bold text-[#666]">
-                Or
-              </span>
-            </div>
-          </>
-        )}
-        <CredentialsLogin />
-      </section>
-    </main>
+    <Auth>
+      <Auth.Image src="/slides/slide-signin.jpg" alt="Sign In" />
+      <Auth.Options>
+        <Auth.OAuthLogin />
+        <Auth.CredentialsLogin />
+      </Auth.Options>
+    </Auth>
   );
 }
