+++ src/app/(auth)/register/page.tsx
@@ -1,9 +1,12 @@
-import { CredentialsRegister } from "@/components/Auth";
+import { Auth } from "@/components/Auth";
 
 export default function Page() {
   return (
-    <main className="grid h-full place-content-center place-items-center">
-      <CredentialsRegister />
-    </main>
+    <Auth>
+      <Auth.Image src="/slides/slide-register.jpg" alt="Sign Up" />
+      <Auth.Options>
+        <Auth.CredentialsRegister />
+      </Auth.Options>
+    </Auth>
   );
 }
