+++ src/components/Auth/components/AuthImage.tsx
@@ -1,0 +1,15 @@
+type AuthImageProps = {
+  src: string;
+  alt: string;
+};
+
+export function AuthImage({ src, alt }: AuthImageProps) {
+  return (
+    <img
+      aria-hidden
+      src={src}
+      alt={alt}
+      className="aspect-[1/1.02] w-full object-cover"
+    />
+  );
+}
