+++ src/components/Auth/components/AuthOptions.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+export function AuthOptions({ children }: { children: ReactNode }) {
+  return (
+    <section className="grid flex-1 place-content-center p-4">
+      {children}
+    </section>
+  );
+}
